Replaces manual protobuf build with Python packages
Manually building protobuf and its Python extension isn't required since Google distributes Python protobuf package as a "wheel" (i.e. a binary compilation). This wheel is assembled with C++ extension, which links libprotobuf statically into it. This way it doesn't depend on having a runtime in the system. With this change it's possible to provision the whole Tyk on-prem stack even on a most basic DigitalOcean instance.

In addition to this, readme instructions are a bit actualised.